requirements.txt
Jinja2==3.1.4
PyYAML==6.0.2
Markdown==3.7